medi_reminder/ai/views.py
"""
Views for the ai app.

This module contains Django views for AI-related functionality
including OCR processing, medication recognition, and AI insights.
"""

from django.shortcuts import render, get_object_or_404
from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
from django.utils.decorators import method_decorator
from django.views import View
from django.contrib.auth.decorators import login_required
from django.core.files.storage import default_storage
from django.core.files.base import ContentFile
from .models import OCRResult, MedicationRecognition, AIInsight
import json
import pytesseract
from PIL import Image


@method_decorator(csrf_exempt, name='dispatch')
class ImageUploadView(View):
    """
    View for uploading and processing images.
    
    Handles POST requests to upload images for OCR and medication recognition.
    """
    def post(self, request):
        # Implementation for image upload and processing
        pass


@method_decorator(csrf_exempt, name='dispatch')
class OCRProcessView(View):
    """
    View for OCR processing.
    
    Handles POST requests to perform OCR on uploaded images.
    """
    def post(self, request):
        # Implementation for OCR processing
        pass


@method_decorator(csrf_exempt, name='dispatch')
class MedicationRecognitionView(View):
    """
    View for medication recognition.
    
    Handles POST requests to recognize medications from OCR text.
    """
    def post(self, request):
        # Implementation for medication recognition
        pass


@login_required
def ai_insights(request):
    """
    View for AI insights.
    
    Returns AI-generated insights and recommendations for the current user.
    """
    # Implementation for AI insights
    return JsonResponse({'insights': []})


@method_decorator(csrf_exempt, name='dispatch')
class OCRResultView(View):
    """
    View for OCR results.
    
    Handles GET requests to retrieve OCR processing results.
    """
    def get(self, request, result_id):
        # Implementation for getting OCR results
        pass


@method_decorator(csrf_exempt, name='dispatch')
class MedicationRecognitionDetailView(View):
    """
    View for medication recognition details.
    
    Handles GET, PUT, DELETE requests for medication recognition results.
    """
    def get(self, request, recognition_id):
        # Implementation for getting recognition details
        pass
    
    def put(self, request, recognition_id):
        # Implementation for updating recognition
        pass
    
    def delete(self, request, recognition_id):
        # Implementation for deleting recognition
        pass


@login_required
def user_ai_history(request):
    """
    View for user's AI processing history.
    
    Returns history of OCR and medication recognition for the current user.
    """
    # Implementation for AI history
    return JsonResponse({'history': []})